% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{compute_observed_statistics}
\alias{compute_observed_statistics}
\title{Compute the test and information for auxiliary statistics}
\usage{
compute_observed_statistics(x, clust1, clust2)
}
\arguments{
\item{x}{n-by-p data matrix}

\item{clust1}{a vector giving a subset of \code{1:n}}

\item{clust2}{a vector giving a subset of \code{1:n}}
}
\description{
This computes the test statistic, which is the ratio of the squared Frobenius
norms of two matrices \code{b} and \code{w} times a scaling factor.  This test
statistic is F-distributed when \code{clust1} and \code{clust2} are predetermined (i.e.
not learned on data being used for test).  But in our setting, the clusters
are learned on the same data.  To properly account for this double dipping,
we will also need the values of some additional statistics, which depend on
\code{b} and \code{w}.
}