% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{compute_pvalue}
\alias{compute_pvalue}
\title{Compute selective p-value}
\usage{
compute_pvalue(t, x, hc, hc_sequence = NULL, obs_statistics = NULL, dist_func)
}
\arguments{
\item{t}{which step p-value is being calculated at}

\item{x}{observed data matrix}

\item{hc}{output of \code{hclust_randomized()}}

\item{hc_sequence}{output of \code{get_hclust_sequence()} applied to output of
\code{hclust_randomized()}}

\item{obs_statistics}{output of \code{compute_observed_statistics()}}

\item{dist_func}{the function that was used to compute the original n by n matrix}
}
\description{
Arguments \code{hc_sequence} and \code{obs_statistics} if \code{NULL} are computed from \code{hc}
and \code{t}.
}