% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{compute_pvalue_grid}
\alias{compute_pvalue_grid}
\title{Compute selective p-value}
\usage{
compute_pvalue_grid(
  t,
  x,
  hc,
  hc_sequence = NULL,
  obs_statistics = NULL,
  dist_func
)
}
\description{
Compute selective p-value
}