% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{compute_selected_probs}
\alias{compute_selected_probs}
\title{Compute selection probabilities on a new data set}
\usage{
compute_selected_probs(x_new, hc_sequence, dist_func, linkage_func, tau)
}
\arguments{
\item{x_new}{observed data matrix}

\item{hc_sequence}{output of \code{get_hclust_sequence()} applied to output of
\code{hclust_randomized()}}

\item{dist_func}{the function that was used to compute the original n by n matrix}

\item{linkage_func}{the linkage used to measure the dissimilarity
between pairs of clusters}
}
\description{
These probabilites are used in the integrand of the p-value calculation.
}