% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{generate_data}
\alias{generate_data}
\title{Generate samples from a clustering model}
\usage{
generate_data(nums, mu, sigma = 1)
}
\arguments{
\item{nums}{k-vector giving the number of points in each cluster}

\item{mu}{k-by-p matrix giving the k cluster centers}

\item{sigma}{within cluster sd}
}
\description{
Samples X|Y=l ~ N_p(mu_l, sigma^2 * I_p)
}