% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{get_hclust_sequence}
\alias{get_hclust_sequence}
\title{Get sequence of clusters and merges from an \code{hclust} object}
\usage{
get_hclust_sequence(hc)
}
\arguments{
\item{hc}{an object of class \code{hclust}}
}
\value{
\code{clusts} is a list, where \code{clusts[[i]]} is the clustering before
merge i occurs.  The first cluster in \code{clusts[[i]]} is the one that was newly
created. \code{merges} is a matrix whose i-th row gives the indices of the
two clusters in \code{clusts[[i]]} being merged
}
\description{
Get sequence of clusters and merges from an \code{hclust} object
}