% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{get_merge_probs}
\alias{get_merge_probs}
\title{Compute merge probabilities}
\usage{
get_merge_probs(d, tau)
}
\arguments{
\item{d}{a symmetric k-by-k matrix of dissimilarities between clusters}

\item{tau}{amount of randomization (tau = 0 is without randomization)}
}
\description{
Computes merge probability vector as in Section 2. When tau is 0, puts equal
probability on all merges that have minimum dissimilarity.  Probabilities are
taken to be 0 on upper triangle and diagonal.
}