% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{hclust_randomized}
\alias{hclust_randomized}
\title{Randomized Hierarchical Clustering}
\usage{
hclust_randomized(d, linkage_func = linkage_complete, tau = 0.1)
}
\arguments{
\item{d}{a symmetric n-by-n matrix of dissimilarities}

\item{linkage_func}{a function taking two index sets in {1,...,n}}

\item{tau}{amount of randomization (tau = 0 is without randomization)}
}
\description{
Performs hierarchical clustering with randomization and outputs an \code{hclust}
object.
}