% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{integrand}
\alias{integrand}
\title{Integrand in p-value calculation}
\usage{
integrand(r, t, x, hc_sequence, obs_statistics, dist_func, linkage_func, tau)
}
\arguments{
\item{r}{scalar value at which we evaluate the integrand}

\item{t}{which step p-value is being calculated at}

\item{x}{observed data matrix}

\item{hc_sequence}{output of \code{get_hclust_sequence()} applied to output of
\code{hclust_randomized()}}

\item{obs_statistics}{output of \code{compute_observed_statistics()}}

\item{dist_func}{the function that was used to compute the original n by n matrix}

\item{linkage_func}{the linkage used to measure the dissimilarity
between pairs of clusters}
}
\description{
Creates a perturbed new data matrix based on the observed data and a
perturbation of size r. Then \code{hc$merge} to reconstruct the possible merges at
each step s from 1 to t. Then, computes the probability of this sequence of
merges being chosen for on this perturbed data set.
}