% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{linkage_complete}
\alias{linkage_complete}
\title{Complete linkage}
\usage{
linkage_complete(clust1, clust2, d)
}
\arguments{
\item{clust1}{a vector giving a subset of \code{1:n}}

\item{clust2}{a vector giving a subset of \code{1:n}}

\item{d}{an n-by-n matrix of dissimilarities}
}
\description{
Complete linkage
}