% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{linkage_group_average}
\alias{linkage_group_average}
\title{Group average linkage}
\usage{
linkage_group_average(clust1, clust2, d)
}
\arguments{
\item{clust1}{a vector giving a subset of \code{1:n}}

\item{clust2}{a vector giving a subset of \code{1:n}}

\item{d}{an n-by-n matrix of dissimilarities}
}
\description{
Group average linkage
}