% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{linkage_single}
\alias{linkage_single}
\title{Single linkage}
\usage{
linkage_single(clust1, clust2, d)
}
\arguments{
\item{clust1}{a vector giving a subset of \code{1:n}}

\item{clust2}{a vector giving a subset of \code{1:n}}

\item{d}{an n-by-n matrix of dissimilarities}
}
\description{
Single linkage
}