% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\name{plot_clustered_data}
\alias{plot_clustered_data}
\title{Plot two-dimensional data while showing up to two clusterings}
\usage{
plot_clustered_data(
  x,
  clust1,
  clust1_label = "clust1",
  clust2 = NULL,
  clust2_label = "clust2"
)
}
\arguments{
\item{x}{n-by-2 data matrix}

\item{clust1}{n-vector of cluster indices}

\item{clust1_label}{name of clustering 1 (e.g. "true clustering")}

\item{clust2}{n-vector of cluster indices (optional)}

\item{clust2_label}{name of clustering 2 (e.g. "estimated clustering"); optional}
}
\description{
For example, can be used to plot both the true and estimated clusterings.
}