% Generated by roxygen2: do not edit by hand
% Please edit documentation in _main.Rmd.
\docType{package}
\name{randhclust-package}
\alias{randhclust}
\alias{randhclust-package}
\title{A Package Defined in a Bookdown}
\description{
This package uses \code{litr} to define an R package through writing a \code{bookdown}.
}
\author{
\strong{Maintainer}: Jacob Bien \email{jbien@usc.edu}

Authors:
\itemize{
  \item Di Wu \email{judydw@umich.edu}
  \item Snigdha Panigrahi \email{psnigdha@umich.edu}
}

}
\keyword{internal}